requirements.txt
# WikiSeed Production Dependencies

# HTTP client
requests>=2.31.0

# Internet Archive
internetarchive>=3.5.0

# Torrent creation
torf>=4.2.0

# Job scheduling
schedule>=1.2.0

# Web framework (Monitor)
Flask>=3.0.0
flasgger>=0.9.7  # Swagger UI for Flask

# Cloudflare R2 (S3-compatible)
boto3>=1.34.0

# YAML configuration
PyYAML>=6.0.0

# Archive.today integration
# (uses requests)

# GitHub Gist
# (uses requests)

# pastebin
# (uses requests)
